requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.10
distlib==0.3.4
Django==4.0.1
django-bootstrap-icons==0.7.5
django-bootstrap-v5==1.0.8
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-oauth==1.1.0
filelock==3.4.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==3.10.1
pipenv==2022.1.8
platformdirs==2.4.1
psycopg2-binary==2.9.3
pytz==2021.3
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.8
virtualenv==20.13.0
virtualenv-clone==0.5.7
zipp==3.7.0
